Fixed "doubleclicking" on Colorpicker
Bug: #315 / #345 / #309
Didn't read all the code so hope i got it right.

When you close a colorchannel it saves the atoms to reuse the gradiets. When opening it again it checks if the old atoms are there and it copies the gradient. But it only checks if the atom is there so when you close the channel before it has drawn all gradients, the gradient is undefined but isGradient is true.

I made it that it also copies the isGradient value from the old atom instead of just setting it true.