multi-squares in multi-squares in ... (#362)
* Added MultiMultiCell support for brush and rules

I changed the splitCellToDiagram and the updatePaddleRule function to go recursivcely through all diagrams. Also empty cells in a diagram get replaced with a cell.

* Merge and split multicells only if somewhat necessary

If a cell on both sides are split the same they only get recoloured.
If they are split the same but they have a different multi-multi cell structure it still merges and splits.
eg: left has a 1x4 multicell and right has 1x2 multi cell with 1x2 multi cells in it. That still doesn't work.

* flatten function and origin of diagram is top left